DueTC Usage
Copy the files DueTC,h and DueTC.cpp in a folder DueTC in your Arduino
IDE library folder.

Use the DueTC_demo.ino to demonstrate the capabilities.  An oscilloscope
is required to monitor the results.  Modify the code and connect one of
the interrupt handlers to toggle pin 13 for visual observations.